Improved summary products services

+++ modules/Vtiger/actions/RelationAjax.php
@@ -362,12 +362,12 @@ class Vtiger_RelationAjax_Action extends \App\Controller\Action
 				throw new \App\Exceptions\NoPermitted('LBL_PERMISSION_DENIED', 406);
 			}
 		}
-		if (!empty($categoryToAdd)) {
+		if (!empty($categoryToAdd) && $relationModel->isTreeRelation()) {
 			foreach ($categoryToAdd as $category) {
 				$relationModel->addRelTree($sourceRecordId, $category);
 			}
 		}
-		if (!empty($categoryToRemove)) {
+		if (!empty($categoryToRemove) && $relationModel->isTreeRelation()) {
 			if ($relationModel->privilegeToDelete()) {
 				foreach ($categoryToRemove as $category) {
 					$relationModel->deleteRelTree($sourceRecordId, $category);
